Extract views into their own components
