+++ fronts-client/src/components/Subnav/index.tsx
@@ -1,40 +1,17 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { Switch, Route, Redirect } from 'react-router-dom';
-import { useHistory, useParams } from 'react-router';
+import { Switch, Route } from 'react-router-dom';
+import { useHistory } from 'react-router';
 import { selectHasSubnavPermission } from 'selectors/configSelectors';
-import {
-	subnavRoutes,
-	subnavListProps,
-	subnavCreateProps,
-	subnavEditProps,
-} from 'routes/routes';
+import { subnavRoutes } from 'routes/routes';
 import { actionAddNotificationBanner } from 'bundles/notificationsBundle';
-import ButtonDefault from 'components/inputs/ButtonDefault';
-import {
-	deleteSubnav,
-	discardSubnav,
-	fetchSubnavConfig,
-	publishSubnav,
-	unpublishSubnav,
-	upsertSubnav,
-} from './subnavApi';
-import SubnavForm from './SubnavForm';
+import { fetchSubnavConfig, upsertSubnav } from './subnavApi';
+import { SubnavFormView } from './SubnavFormView';
+import { SubnavEditRoute } from './SubnavEditRoute';
+import { SubnavListView } from './SubnavListView';
+import { toListEntries } from './helpers';
 import { CustomSubnav, CustomSubnavConfig } from './types';
-import {
-	BackButton,
-	EditStatusBar,
-	List,
-	ListHeader,
-	ListItem,
-	ListItemActions,
-	ListItemMeta,
-	ListItemTitle,
-	Message,
-	StatusTag,
-	SubnavContainer,
-	SubnavContainerHeading,
-} from './styles';
+import { Message, SubnavContainer, SubnavContainerHeading } from './styles';
 
 const NoPermission = () => (
 	<SubnavContainer>
@@ -46,331 +23,6 @@ const NoPermission = () => (
 		</Message>
 	</SubnavContainer>
 );
-
-interface SubnavListEntry {
-	id: string;
-	/** The version shown/edited: the draft if present, otherwise the live copy. */
-	subnav: CustomSubnav;
-	hasLive: boolean;
-	hasDraft: boolean;
-}
-
-/**
- * Build a de-duplicated list of subnavs for display. A subnav can exist in
- * `draft`, `live`, or both; each id appears once, preferring the draft copy.
- */
-const toListEntries = (config: CustomSubnavConfig): SubnavListEntry[] => {
-	const liveById = new Map(config.live.map((s) => [s.id, s]));
-	const draftById = new Map(config.draft.map((s) => [s.id, s]));
-	const ids = [
-		...config.draft.map((s) => s.id),
-		...config.live.filter((s) => !draftById.has(s.id)).map((s) => s.id),
-	];
-	return ids.map((id) => {
-		const draft = draftById.get(id);
-		const live = liveById.get(id);
-		return {
-			id,
-			subnav: (draft ?? live) as CustomSubnav,
-			hasLive: live !== undefined,
-			hasDraft: draft !== undefined,
-		};
-	});
-};
-
-// The version shown when editing: the draft if present, otherwise the live copy.
-const findSubnav = (
-	config: CustomSubnavConfig,
-	id: string,
-): CustomSubnav | undefined =>
-	config.draft.find((s) => s.id === id) ?? config.live.find((s) => s.id === id);
-
-type RunAction = (
-	id: string,
-	action: (id: string) => Promise<CustomSubnavConfig>,
-	confirmMessage?: string,
-) => void;
-
-interface SubnavStatusActionsProps {
-	id: string;
-	hasLive: boolean;
-	hasDraft: boolean;
-	isBusy: boolean;
-	runAction: RunAction;
-}
-
-/**
- * Publish / discard / unpublish / delete buttons for a single subnav, shared by
- * the list rows and the individual edit page. Which buttons show depends on
- * whether the subnav currently has a live and/or draft version.
- */
-const SubnavStatusActions = ({
-	id,
-	hasLive,
-	hasDraft,
-	isBusy,
-	runAction,
-}: SubnavStatusActionsProps) => (
-	<>
-		{hasDraft && (
-			<ButtonDefault
-				type="button"
-				size="s"
-				priority="primary"
-				disabled={isBusy}
-				onClick={() =>
-					runAction(
-						id,
-						publishSubnav,
-						'Publish this subnav? It will go live on the targeted pages.',
-					)
-				}
-			>
-				Publish
-			</ButtonDefault>
-		)}
-		{hasDraft && hasLive && (
-			<ButtonDefault
-				type="button"
-				size="s"
-				priority="muted"
-				disabled={isBusy}
-				onClick={() =>
-					runAction(
-						id,
-						discardSubnav,
-						'Discard draft changes and revert to the live version?',
-					)
-				}
-			>
-				Discard changes
-			</ButtonDefault>
-		)}
-		{hasLive && (
-			<ButtonDefault
-				type="button"
-				size="s"
-				priority="muted"
-				disabled={isBusy}
-				onClick={() =>
-					runAction(
-						id,
-						unpublishSubnav,
-						hasDraft
-							? 'Take this subnav down? You have draft changes — taking it down will undo them.'
-							: 'Take this subnav down? It will no longer show on the targeted pages (kept as a draft).',
-					)
-				}
-			>
-				Unpublish
-			</ButtonDefault>
-		)}
-		<ButtonDefault
-			type="button"
-			size="s"
-			priority="muted"
-			disabled={isBusy}
-			onClick={() =>
-				runAction(
-					id,
-					deleteSubnav,
-					'Delete this subnav entirely? This removes both the live and draft versions.',
-				)
-			}
-		>
-			Delete
-		</ButtonDefault>
-	</>
-);
-
-const SubnavStatusTags = ({
-	hasLive,
-	hasDraft,
-}: {
-	hasLive: boolean;
-	hasDraft: boolean;
-}) => (
-	<>
-		{hasLive && <StatusTag>Live</StatusTag>}
-		{hasDraft && (
-			<StatusTag draft>{hasLive ? 'Draft changes' : 'New draft'}</StatusTag>
-		)}
-	</>
-);
-
-interface SubnavListViewProps {
-	entries: SubnavListEntry[];
-	isLoading: boolean;
-	pendingActionId: string | null;
-	onCreate: () => void;
-	onEdit: (id: string) => void;
-	runAction: (
-		id: string,
-		action: (id: string) => Promise<CustomSubnavConfig>,
-		confirmMessage?: string,
-	) => void;
-}
-
-const SubnavListView = ({
-	entries,
-	isLoading,
-	pendingActionId,
-	onCreate,
-	onEdit,
-	runAction,
-}: SubnavListViewProps) => (
-	<SubnavContainer>
-		<SubnavContainerHeading>Custom subnavs</SubnavContainerHeading>
-
-		<ListHeader>
-			<ButtonDefault type="button" priority="primary" onClick={onCreate}>
-				+ Create new subnav
-			</ButtonDefault>
-		</ListHeader>
-
-		{isLoading ? (
-			<Message>Loading…</Message>
-		) : entries.length === 0 ? (
-			<Message>No custom subnavs yet. Create one to get started.</Message>
-		) : (
-			<List>
-				{entries.map(({ id, subnav, hasLive, hasDraft }) => {
-					const isBusy = pendingActionId === id;
-					return (
-						<ListItem key={id}>
-							<div>
-								<ListItemTitle>
-									{subnav.header.headerText || 'Untitled subnav'}
-									<SubnavStatusTags hasLive={hasLive} hasDraft={hasDraft} />
-								</ListItemTitle>
-								<ListItemMeta>
-									{subnav.pages.length} page
-									{subnav.pages.length === 1 ? '' : 's'} · {subnav.links.length}{' '}
-									link
-									{subnav.links.length === 1 ? '' : 's'}
-								</ListItemMeta>
-							</div>
-							<ListItemActions>
-								<ButtonDefault
-									type="button"
-									size="s"
-									disabled={isBusy}
-									onClick={() => onEdit(id)}
-								>
-									Edit
-								</ButtonDefault>
-								<SubnavStatusActions
-									id={id}
-									hasLive={hasLive}
-									hasDraft={hasDraft}
-									isBusy={isBusy}
-									runAction={runAction}
-								/>
-							</ListItemActions>
-						</ListItem>
-					);
-				})}
-			</List>
-		)}
-	</SubnavContainer>
-);
-
-interface SubnavFormViewProps {
-	heading: string;
-	statusBar?: React.ReactNode;
-	initialSubnav?: CustomSubnav;
-	onSave: (subnav: CustomSubnav) => Promise<void>;
-	onCancel: () => void;
-	saving: boolean;
-}
-
-const SubnavFormView = ({
-	heading,
-	statusBar,
-	initialSubnav,
-	onSave,
-	onCancel,
-	saving,
-}: SubnavFormViewProps) => (
-	<SubnavContainer>
-		<BackButton type="button" onClick={onCancel}>
-			← Back to subnavs
-		</BackButton>
-		<SubnavContainerHeading>{heading}</SubnavContainerHeading>
-		{statusBar}
-		<SubnavForm initialSubnav={initialSubnav} onSave={onSave} saving={saving} />
-	</SubnavContainer>
-);
-
-interface SubnavEditRouteProps {
-	config: CustomSubnavConfig | null;
-	isLoading: boolean;
-	pendingActionId: string | null;
-	runAction: RunAction;
-	onSave: (subnav: CustomSubnav) => Promise<void>;
-	onCancel: () => void;
-	saving: boolean;
-}
-
-/**
- * Resolves the `:id` route param against the loaded config. While the config is
- * still loading we show a placeholder; if the id is unknown (e.g. deleted or a
- * stale link) we send the user back to the list.
- */
-const SubnavEditRoute = ({
-	config,
-	isLoading,
-	pendingActionId,
-	runAction,
-	onSave,
-	onCancel,
-	saving,
-}: SubnavEditRouteProps) => {
-	const { id } = useParams<{ id: string }>();
-
-	if (isLoading || !config) {
-		return (
-			<SubnavContainer>
-				<SubnavContainerHeading>Edit custom subnav</SubnavContainerHeading>
-				<Message>Loading…</Message>
-			</SubnavContainer>
-		);
-	}
-
-	const subnav = findSubnav(config, id);
-	if (!subnav) {
-		return <Redirect to={subnavRoutes.base} />;
-	}
-
-	const hasLive = config.live.some((s) => s.id === id);
-	const hasDraft = config.draft.some((s) => s.id === id);
-
-	return (
-		<SubnavFormView
-			heading="Edit custom subnav"
-			statusBar={
-				<EditStatusBar>
-					<div>
-						<SubnavStatusTags hasLive={hasLive} hasDraft={hasDraft} />
-					</div>
-					<ListItemActions>
-						<SubnavStatusActions
-							id={id}
-							hasLive={hasLive}
-							hasDraft={hasDraft}
-							isBusy={pendingActionId === id}
-							runAction={runAction}
-						/>
-					</ListItemActions>
-				</EditStatusBar>
-			}
-			initialSubnav={subnav}
-			onSave={onSave}
-			onCancel={onCancel}
-			saving={saving}
-		/>
-	);
-};
 
 const SubnavSection = () => {
 	const hasPermission = useSelector(selectHasSubnavPermission);
@@ -461,7 +113,7 @@ const SubnavSection = () => {
 
 	return (
 		<Switch>
-			<Route {...subnavCreateProps}>
+			<Route {...subnavRoutes.createProps}>
 				<SubnavFormView
 					heading="Create custom subnav"
 					onSave={handleCreate}
@@ -469,7 +121,7 @@ const SubnavSection = () => {
 					saving={isSaving}
 				/>
 			</Route>
-			<Route {...subnavEditProps}>
+			<Route {...subnavRoutes.editProps}>
 				<SubnavEditRoute
 					config={subnavConfig}
 					isLoading={isLoading}
@@ -480,7 +132,7 @@ const SubnavSection = () => {
 					saving={isSaving}
 				/>
 			</Route>
-			<Route {...subnavListProps}>
+			<Route {...subnavRoutes.listProps}>
 				<SubnavListView
 					entries={entries}
 					isLoading={isLoading}
